Move root_hash() to proof_list_index module
It makes more sense to keep this utility function close to
ProofListIndex. It is effectively an optimized way to compute the same
root hash as available via ProofListIndex.

Remove now unnecessary helpers::merkle module. All Merkle-related code
is now located in the "storage" module.
